refactor(agent): extract Happy-archive seam from agent.store (#3250)
First incremental, behavior-preserving extraction from the 9.5k-line
agent.store.ts god-file. Moves the pure Happy-origin archive seam —
HappyArchiveFence, planHappyArchiveInvalidation,
planHappyProviderArchiveInvalidation, retireHappyArchivedSiblingProvider,
and the HappyArchivedSiblingRetirementResult type — into a co-located
module at src/lib/agent/happy-archive.ts.

These are pure functions/classes with no closure over store state, so the
move is a straight relocation. The store keeps the happyArchiveFence
singleton and every call site unchanged, and re-exports the moved symbols
so the public surface of @/stores/agent.store is identical. agent.store.ts
drops from 9644 to 9474 lines.

Establishes the co-location pattern the issue calls for so subsequent
seams can follow. No behavior, IPC-contract, or persistence-schema change.

Validation: full unit suite green (2714 passed), agent-history-persistence
and agent-conversation-archive-invalidation pinned tests unchanged and
passing, Biome clean, both files type-clean under tsc.

Taariq Lewis, SerenAI, Paloma, and Volume at https://serendb.com
Email: hello@serendb.com
